# backend/openapi.json
{"openapi":"3.0.0","paths":{"/api/content":{"get":{"operationId":"ContentController_getAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Chapter"}}}}}},"security":[{"bearer":[]}]}},"/api/user/info":{"get":{"operationId":"UserController_getInfo","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"security":[{"bearer":[]}]}},"/api/user/onboard":{"post":{"operationId":"UserController_processOnboarding","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OnboardingDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}]}},"/api/auth/verify":{"post":{"operationId":"AuthController_verify","summary":"","description":"Given a [Firebase TokenId](https://firebase.google.com/docs/auth/admin/verify-id-tokens#android), verify the token and return\nanother a JWT to be used for future API calls to protected endpoints (in Bearer authentication).","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequestDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}}}},"/api/healer/nearby":{"get":{"operationId":"HealerController_nearby","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NearbyHealersDto"}}}}},"security":[{"bearer":[]}]}},"/api/healer/availability":{"get":{"operationId":"HealerController_availability","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AvailabilitySlotDto"}}}}}},"security":[{"bearer":[]}]}},"/api/healer/book":{"post":{"operationId":"HealerController_book","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}]}}},"info":{"title":"HEA Backend","description":"See /api/auth/verify for authentication","version":"0.5","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"Chapter":{"type":"object","properties":{"id":{"type":"number"},"chapterNum":{"type":"number"},"icon":{"type":"string"},"title":{"type":"string"},"subchapters":{"type":"array","items":{"$ref":"#/components/schemas/Subchapter"}}},"required":["id","chapterNum","icon","title","subchapters"]},"Page":{"type":"object","properties":{"id":{"type":"number"},"pageNum":{"type":"number"},"icon":{"type":"string"},"text":{"type":"string"},"subchapter":{"$ref":"#/components/schemas/Subchapter"}},"required":["id","pageNum","icon","text","subchapter"]},"Subchapter":{"type":"object","properties":{"id":{"type":"number"},"subchapterNum":{"type":"number"},"icon":{"type":"string"},"title":{"type":"string"},"callToAction":{"type":"string"},"chapter":{"$ref":"#/components/schemas/Chapter"},"pages":{"type":"array","items":{"$ref":"#/components/schemas/Page"}}},"required":["id","subchapterNum","icon","title","callToAction","chapter","pages"]},"User":{"type":"object","properties":{"id":{"type":"number"},"authId":{"type":"string"},"level":{"type":"object"},"healthData":{"type":"array","items":{"type":"object"}},"name":{"type":"string"},"gender":{"type":"string","enum":["Male","Female","Others"]},"birthday":{"format":"date-time","type":"string"},"height":{"type":"number"},"weight":{"type":"number"},"country":{"type":"string"},"isSmoker":{"type":"boolean"},"smokingPacksPerDay":{"type":"string","enum":["LessOnePack","OneToTwoPacks","ThreeToFivePacks","FivePacks"]},"smokingYears":{"type":"number"},"alcoholFreq":{"type":"string","enum":["NotAtAll","OnceAMonth","OnceAWeek","FewTimesAWeek","Everyday"]},"outlook":{"type":"string","enum":["Positive","Negative","Soulless"]},"maritalStatus":{"type":"string","enum":["Single","Married","Widowed","Divorced"]},"familyHistory":{"type":"string"},"birthControl":{"type":"string"}},"required":["id","authId","level","healthData"]},"SmokerInfo":{"type":"object","properties":{"packsPerDay":{"type":"string","enum":["LessOnePack","OneToTwoPacks","ThreeToFivePacks","FivePacks"]},"years":{"type":"number","minimum":0}},"required":["packsPerDay","years"]},"OnboardingDto":{"type":"object","properties":{"name":{"type":"string"},"gender":{"type":"string","enum":["Male","Female","Others"]},"birthday":{"format":"date-time","type":"string","description":"Birthday, but only the Date part is considered"},"height":{"type":"number","minimum":0.3,"maximum":3},"weight":{"type":"number","minimum":3,"maximum":200},"country":{"type":"string","description":"ISO31661 2-letter alphabetical country code"},"smoking":{"nullable":true,"allOf":[{"$ref":"#/components/schemas/SmokerInfo"}]},"alcoholFreq":{"type":"string","enum":["NotAtAll","OnceAMonth","OnceAWeek","FewTimesAWeek","Everyday"]},"outlook":{"type":"string","enum":["Positive","Negative","Soulless"]},"maritalStatus":{"type":"string","enum":["Single","Married","Widowed","Divorced"]},"familyHistory":{"type":"string","nullable":true},"birthControl":{"type":"string","nullable":true}},"required":["name","gender","birthday","height","weight","country","smoking","alcoholFreq","outlook","maritalStatus","familyHistory","birthControl"]},"AuthRequestDto":{"type":"object","properties":{"firebaseToken":{"type":"string"}},"required":["firebaseToken"]},"AuthResponseDto":{"type":"object","properties":{"jwt":{"type":"string"}},"required":["jwt"]},"MedicalTagDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"proficiency":{"type":"number"}},"required":["name","description","proficiency"]},"AvailabilitySlotDto":{"type":"object","properties":{"start":{"format":"date-time","type":"string"},"end":{"format":"date-time","type":"string"}},"required":["start","end"]},"NearbyHealerDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"description":{"type":"string"},"tags":{"type":"array","items":{"$ref":"#/components/schemas/MedicalTagDto"}},"availability":{"type":"array","items":{"$ref":"#/components/schemas/AvailabilitySlotDto"}},"rocation_lat":{"type":"number"},"location_lng":{"type":"number"}},"required":["id","name","description","tags","availability","rocation_lat","location_lng"]},"NearbyHealersDto":{"type":"object","properties":{"healers":{"type":"array","items":{"$ref":"#/components/schemas/NearbyHealerDto"}}},"required":["healers"]},"BookingDto":{"type":"object","properties":{"healerId":{"type":"number"},"slot":{"$ref":"#/components/schemas/AvailabilitySlotDto"}},"required":["healerId","slot"]}}}}